# packages/ts/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "AudioPreviewRequest": {
        "additionalProperties": false,
        "properties": {
          "agent_id": {
            "format": "uuid",
            "title": "Agent Id",
            "type": "string"
          },
          "confirmed": {
            "const": true,
            "title": "Confirmed",
            "type": "boolean"
          },
          "text": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "version_id": {
            "format": "uuid",
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": ["agent_id", "version_id", "text", "confirmed"],
        "title": "AudioPreviewRequest",
        "type": "object"
      },
      "AudioPreviewResult": {
        "properties": {
          "agent_id": {
            "format": "uuid",
            "title": "Agent Id",
            "type": "string"
          },
          "audio_base64": {
            "title": "Audio Base64",
            "type": "string"
          },
          "canonical_text": {
            "title": "Canonical Text",
            "type": "string"
          },
          "duration_seconds": {
            "title": "Duration Seconds",
            "type": "number"
          },
          "evaluation_kind": {
            "const": "paid_tts_preview",
            "default": "paid_tts_preview",
            "title": "Evaluation Kind",
            "type": "string"
          },
          "expires_in_seconds": {
            "const": 60,
            "default": 60,
            "title": "Expires In Seconds",
            "type": "integer"
          },
          "media_type": {
            "const": "audio/wav",
            "default": "audio/wav",
            "title": "Media Type",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "speech_normalized_text": {
            "title": "Speech Normalized Text",
            "type": "string"
          },
          "version_id": {
            "format": "uuid",
            "title": "Version Id",
            "type": "string"
          },
          "voice": {
            "title": "Voice",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "agent_id",
          "version_id",
          "audio_base64",
          "provider",
          "model",
          "voice",
          "duration_seconds",
          "canonical_text",
          "speech_normalized_text"
        ],
        "title": "AudioPreviewResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Or-On Platform Control API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/orchestration/agents/audio-preview": {
      "post": {
        "operationId": "preview_agent_audio",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioPreviewResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Preview Agent Audio",
        "tags": ["orchestration"]
      }
    },
...
  }
}